pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "swaggertest"
version = "1.0.0"
description = "A tool that discovers OpenAPI specs from Swagger UI, tests GET endpoints, and produces structured JSON reports."
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "pyyaml>=6.0",
    "prance>=23.6",
    "openapi-spec-validator>=0.7",
    "jsonschema>=4.20",
    "typer>=0.12",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
]

[project.scripts]
swaggertest = "swaggertest.cli:app"

[tool.setuptools.packages.find]
include = ["swaggertest*"]
